// src/typeScript/constants/enum.ts
export enum FORM {
  LOGIN = "login",
  RESGITER = "register",

  EMAIL = "email",
  PASSWORD = "password",
  CONFIRMPASSWORD = "confirm-password",
}

export enum FORM_EVEN {
  INPUT = "input",
  BLUR = "blur",
  FOCUS = "focus",
}

export enum KEY {
  LOCALSTORAGE_UESR = "user",
  LOCALSTORAGE_TODOS = "todos",
  LOCALSTORAGE_ID_UPDATE = "idUpdate",
}
export enum TOAST {
  SUCCESS = "success",
  ERROR = "error",
}

export enum PAGE {
  LOGIN = "loginPage",
  TODO = "todoPage",
}

export enum ACTION_FORM {
  ADD = "ADD",
  UPDATE = "UPDATE",
}